+++ docs/reference/javadoc/package-search-index.js
@@ -1,1 +1,0 @@
-packageSearchIndex = [{"l":"All Packages","url":"allpackages-index.html"},{"l":"com.caoccao.javet.javenode"},{"l":"com.caoccao.javet.javenode.enums"},{"l":"com.caoccao.javet.javenode.interfaces"},{"l":"com.caoccao.javet.javenode.modules"},{"l":"com.caoccao.javet.javenode.modules.timers"}]